Remove spring-boot plugin, add application and flyway plugins
We're not really using spring-boot to any advantage. It lets us omit
version numbers, but we don't want to do that anyway.

Adding the application plugin will let us run the app through gradle,
and automatically pulls in the distribution plugin to create zip and tar
archives.

The flyway plugin will let us version our database via migration
scripts. We'll have to configure Hibernate/Spring to not set up the
database themselves, and we'll have to make sure our migrations are
cross-platform enough to work on H2 and PostgreSQL.