Add Hibernate validator and create separate Spring config
We're going to use the Hibernate validator library to validate our model. We've
created a bean for the validator factory. It looks like Spring also has some
validation, but the docs talk about property editors and horrible things like
that. Hibernate sounds pretty simple.

We've also split out the configuration into its own class, which is imported by
the application.